budget
added party budget

+++ plan_your_event_mobile/lib/models/bugdet_item_model.dart
@@ -2,7 +2,7 @@
 class BudgetItem {
   String id;
   String budgetItem;
-  int price;
+  double price;
   String partyId;
   String budgetId;
